DCAC-222: Rationalise consumers, consumer packages.
