+++ src/main/java/uk/gov/companieshouse/ordernotification/consumer/orderreceived/OrderReceivedConsumer.java
@@ -1,9 +1,8 @@
-package uk.gov.companieshouse.ordernotification.ordersconsumer;
+package uk.gov.companieshouse.ordernotification.consumer.orderreceived;
 
 import org.springframework.kafka.annotation.KafkaListener;
 import org.springframework.messaging.Message;
 import org.springframework.stereotype.Service;
-import uk.gov.companieshouse.itemgroupprocessedsend.ItemGroupProcessedSend;
 import uk.gov.companieshouse.orders.OrderReceived;
 
 /**
@@ -11,15 +10,12 @@ import uk.gov.companieshouse.orders.OrderReceived;
  * ready to be processed.</p>
  */
 @Service
-public class OrdersKafkaConsumer {
+public class OrderReceivedConsumer {
 
     private final OrderMessageHandler orderMessageHandler;
 
-    private final ItemGroupProcessedSendHandler itemGroupProcessedSendHandler;
-
-    public OrdersKafkaConsumer(OrderMessageHandler orderMessageHandler, ItemGroupProcessedSendHandler itemGroupProcessedSendHandler) {
+    public OrderReceivedConsumer(OrderMessageHandler orderMessageHandler) {
         this.orderMessageHandler = orderMessageHandler;
-        this.itemGroupProcessedSendHandler = itemGroupProcessedSendHandler;
     }
 
     /**
@@ -36,23 +32,6 @@ public class OrdersKafkaConsumer {
             containerFactory = "kafkaOrderReceivedListenerContainerFactory")
     public void processOrderReceived(Message<OrderReceived> message) {
         orderMessageHandler.handleMessage(message);
-    }
-
-    /**
-     * <p>Consumes a message from the item-group-processed-send topic</p>
-     *
-     * @param message A {@link Message message} containing an
-     *                {@link ItemGroupProcessedSend entity}.
-     */
-    @KafkaListener(id = "#{'${kafka.topics.item-group-processed-send-group}'}",
-            groupId = "#{'${kafka.topics.item-group-processed-send-group}'}",
-            topics = "#{'${kafka.topics.item-group-processed-send}'}",
-            autoStartup = "#{!${uk.gov.companieshouse.order-notification-sender.error-consumer}}",
-            containerFactory = "kafkaItemGroupProcessedSendListenerContainerFactory")
-    public void processItemGroupProcessedSend(Message<ItemGroupProcessedSend> message) {
-
-        //TODO: Add a RetryTopic annotation?
-        itemGroupProcessedSendHandler.handleMessage(message);
     }
 
     /**
