+++ src/test/java/uk/gov/companieshouse/ordernotification/consumer/orderreceived/OrderMessageHandlerTest.java
@@ -1,4 +1,4 @@
-package uk.gov.companieshouse.ordernotification.ordersconsumer;
+package uk.gov.companieshouse.ordernotification.consumer.orderreceived;
 
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
