+++ src/test/java/uk/gov/companieshouse/ordernotification/consumer/orderreceived/OrderReceivedLoggerTest.java
@@ -1,4 +1,4 @@
-package uk.gov.companieshouse.ordernotification.ordersconsumer;
+package uk.gov.companieshouse.ordernotification.consumer.orderreceived;
 
 import static org.mockito.ArgumentMatchers.anyMap;
 import static org.mockito.ArgumentMatchers.eq;
