+++ src/test/java/uk/gov/companieshouse/ordernotification/emailsender/KafkaProducerConfigurationTest.java
@@ -8,7 +8,7 @@ import org.mockito.junit.jupiter.MockitoExtension;
 import uk.gov.companieshouse.kafka.exceptions.ProducerConfigException;
 import uk.gov.companieshouse.kafka.producer.Acks;
 import uk.gov.companieshouse.kafka.producer.ProducerConfig;
-import uk.gov.companieshouse.ordernotification.ordersconsumer.KafkaConfig;
+import uk.gov.companieshouse.ordernotification.config.KafkaConfig;
 
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.is;
